applied google java style guide
