+++ src/main/resources/application.properties
@@ -7,9 +7,8 @@ spring.application.name=Military Logistics Management System
 spring.datasource.url=jdbc:postgresql://localhost:5432/test_db
 spring.datasource.username=db_admin
 
-# For development, DB Password set up as a user environment variable
-# in the run/debug configuration settings. In Production, using spring cloud config
-# or a secrets management tool would be the better policy.
+# For development, DB Password set up as a user environment variable.
+# In Production, using spring cloud config or a secrets management tool would be the better policy.
 spring.datasource.password=${DB_PASSWORD}
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.hibernate.ddl-auto=create
