+++ src/test/java/com/example/militarylogisticsmgmt/MilitaryLogisticsManagementSystemApplicationTests.java
@@ -1,6 +1,10 @@
 package com.example.militarylogisticsmgmt;
 
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+
+import com.example.militarylogisticsmgmt.controller.UserController;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 
@@ -11,14 +15,31 @@ import org.springframework.test.context.ActiveProfiles;
  * to verify that the application starts up correctly within the test environment.
  * </p>
  *
- *
- * <p>It uses the "test" profile to configure the environment for testing purposes.
+ * <p>{@code @ActiveProfiles("test")} activates the "test" profile, loading properties from
+ * the application-test.properties file.
  * </p>
  *
  */
 @SpringBootTest
 @ActiveProfiles("test")
 class MilitaryLogisticsManagementSystemApplicationTests {
+
+  private final UserController controller;
+
+  /**
+   * Constructor for {@link MilitaryLogisticsManagementSystemApplicationTests}.
+   *
+   * <p>This constructor allows Spring to inject the necessary {@link UserController} bean into
+   * the test class, ensuring that the controller is available for testing and that all its
+   * dependencies are correctly resolved.
+   * </p>
+   *
+   * @param controller the {@link UserController} to be injected, which handles user operations.
+   */
+  @Autowired
+  public MilitaryLogisticsManagementSystemApplicationTests(UserController controller) {
+    this.controller = controller;
+  }
 
   /**
    * Verifies that the Spring application context loads successfully.
@@ -29,6 +50,6 @@ class MilitaryLogisticsManagementSystemApplicationTests {
    */
   @Test
   void contextLoads() {
+    assertThat(controller).isNotNull();
   }
-
 }
