+++ src/test/resources/application-test.properties
@@ -2,14 +2,15 @@
 server.error.include-message=always
 server.error.include-binding-errors=always
 
-# H2 Database
-spring.h2.console.enabled=true
-spring.datasource.url=jdbc:h2:mem:testdb
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=password
-spring.datasource.initialization-mode=always
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+# PostgreSQL Database
+spring.application.name=Military Logistics Management System
+spring.datasource.url=jdbc:postgresql://localhost:5432/test_db
+spring.datasource.username=db_admin
+
+# For development, DB Password set up as a user environment variable.
+# In Production, using spring cloud config or a secrets management tool would be the better policy.
+spring.datasource.password=${DB_PASSWORD}
+spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.hibernate.ddl-auto=create-drop
 
 # Disable Spring Cloud Config for testing
